feat: Align with Mbed-TLS agent
While the previous version of the OTA agent was
using OpenSSL, the updated is using Mbed-TLS. A
problem occurred with the new agent is that the
return value of `mbedtls_ssl_read()` when agent
is done sending the data is:
`MBEDTLS_ERR_SSL_PEER_CLOSE_NOTIFY` while in the
past it used to be zero. The code is updated to
check the new case, too.

Furthermore, a `free` command is removed to avoid
using it, when trying to reconnect.

Signed-off-by: Ilias Lagomatis <ilago@nubificus.co.uk>