feat: Introduce DICE cache
Instead of regenerating the certificate each time
we get an onboard/update request, we use a cache
buffer to hold it there and provide a pointer to
the user. The certificate is generated only the
first time.

FIX: Find an mprotect-like esp32 tool to set the
certificate cache memory as read-only. Currently,
someone could corrupt the certificate buffer.

Signed-off-by: Ilias Lagomatis <ilago@nubificus.co.uk>